==> 02-Collaborative Filtering Recommender System.ipynb <==
# Collaborative Filtering Recommender System


For this part, you can use the python library Scikit-Surprise. Please find the
documentation here: https://surprise.readthedocs.io/en/stable/. However,
you are free to use any libraries of your choice.


# Load cleaned data splits
import pandas as pd 
train_data = pd.read_pickle("train_dataframe.pkl") 
test_data = pd.read_pickle("test_dataframe.pkl")

## Step 1

Based on the frequency of the most rated items computed in Week 6,
implement the TopPop recommender system, which always recommends
the same top-k items sorted decreasingly by the number of “high” ratings
(e.g., ≥ 3) in the training split, train.tsv.


# get top 10 recommendations from TopPop recommender system for next week
def TopPop(train_data, items_rating_count, k):
    high_ratings = train_data[train_data['rating'] >= 3]
    stats = high_ratings.groupby('item_id').agg(num_ratings=('rating', 'count'), avg_rating=('rating', 'mean')).reset_index()
    items_rating_count = stats.sort_values(by=['avg_rating', 'num_ratings'], ascending=False)
    return items_rating_count.head(k)


top_k = 10 
items_rating_count = train_data['item_id'].value_counts()
top_items = TopPop(train_data, items_rating_count, top_k)

print(top_items)
# save it for next week comparison with knn ,svd
top_items.to_csv('top_items.csv', index=False)


## Step 2
Choose at least one neighborhood-based model and one latent factor model
that uses the observed user-item ratings in the training set to predict the
unobserved ratings. Report your choice of models.


### Neighborhood-based model(KNN, K=5)

import random
import pandas as pd
import numpy as np
from surprise import Reader
from surprise import Dataset
from surprise import KNNWithMeans
from surprise.model_selection import KFold
from sklearn.metrics import mean_squared_error as mse
from surprise import accuracy

# Convert train data format
reader = Reader(rating_scale=(1, 5))
training_matrix = Dataset.load_from_df(train_data[['user_id', 'item_id', 'rating']], reader)

# test the implemented KNN models with some randomly selected parameters
sim_options = {'name': 'cosine',
               'user_based': True
               }
algo_knn = KNNWithMeans(k= 5, 
                    random_state=0,
                    sim_options= sim_options,
                    verbose=False)

trainset = training_matrix.build_full_trainset()# includes the entire dataset for training
algo_knn.fit(trainset)

unobserved_ratings = trainset.build_anti_testset()
pred_KNN = algo_knn.test(unobserved_ratings)
test_users = set(test_data['user_id'].unique())
unobserved_users = set(user_id for user_id, item_id, _ in unobserved_ratings)
num_users_not_in_test = len(unobserved_users - test_users)# number of the users that are not in the test set

predictions_new = [pred for pred in pred_KNN if pred.uid in test_users] 
num_predictions = len(predictions_new)
average_prediction = round(np.mean([pred.est for pred in predictions_new]), 3)


print("Number of predictions: ", num_predictions)
print("Average of predictions: ", average_prediction)

### Latent Factor Model(SVD)

# test the implemented SVD models with some randomly selected parameters
from surprise import SVD

algo_SVD = SVD(n_factors=30, n_epochs=100)
algo_SVD.fit(trainset)

pred_SVD = algo_SVD.test(unobserved_ratings)
predictions = [pred for pred in pred_SVD if pred.uid in test_users]

average_preds = round(np.mean([pred.est for pred in predictions]), 3)

print("Number of predictions:", len(predictions))
print("Average of predictions:", average_preds)

## Step 3
Use 5-fold cross-validation on the training set to tune the hyperparameters
of the chosen models (similarity measure and number of neighbors for the
neighborhood-based model; number of latent factors and number of epochs
for the latent factor model).


Choose an evaluation measure that is suitable for this task and justify your
motivation in using it. Report the optimal hyperparameters together with
the scores of your chosen measure, averaged over the 5 folds.


### KNN Model tune+ RMSE as evaluation measure

# the random seed
my_seed = 0
random.seed(my_seed)
np.random.seed(my_seed)

# Define a cross-validation iterator
kf = KFold(n_splits=5)  #5 fold

rmse_result = dict()

# Parameter options
list_neighbour =  range(1, 11)
similarity_measures = ["cosine", "pearson", "msd"]

# test parameter sets by turns
for neighbour in list_neighbour:
    rmse_result[neighbour] = {}

    for sim_measure in similarity_measures:
        rmse_result[neighbour][sim_measure] = {}
        
        fold = 0
        for trainset, testset in kf.split(training_matrix):
            algo = KNNWithMeans(k=neighbour,
                                sim_options={"name": sim_measure, "user_based": True},
                                verbose=False)

            algo.fit(trainset)
            predictions_KNN = algo.test(testset)

            rmse = accuracy.rmse(predictions_KNN, verbose=False)
            rmse_result[neighbour][sim_measure][fold] = rmse

            fold += 1


# show and find the optimal set of parameters by selecting the lowest mean RMSE
rmse_df = pd.DataFrame.from_dict(
    {f"{n}_{m}": v for n, sim_res in rmse_result.items() for m, v in sim_res.items()},
    orient='index')

rmse_df['mean_rmse'] = rmse_df.mean(axis=1)

print(rmse_df[:30])

best_params = rmse_df['mean_rmse'].idxmin()
best_k, best_sim = best_params.split('_')
print()
print(f'The best number of neighbors is: {best_k} with similarity measure: {best_sim} ')


### SVD Model tune + RMSE as evaluation measure

kf = KFold(n_splits=5)  #5 fold
rmse_result = dict()

# parameter options
list_factor = [5, 10, 20, 30, 50]
list_epoch = [10, 20, 30, 50, 100]

# test parameter sets by turns
for factor in list_factor:
    rmse_result[factor] = {}

    for epoch in list_epoch:
        rmse_result[factor][epoch] = {}
        
        fold = 0
        for trainset, testset in kf.split(training_matrix):
            algo = SVD(n_factors=factor, 
                       n_epochs=epoch)

            algo.fit(trainset)
            predictions_SVD = algo.test(testset)

            rmse = accuracy.rmse(predictions_SVD, verbose=False)
            rmse_result[factor][epoch][fold] = rmse

            fold += 1

rmse_df = pd.DataFrame.from_dict(
    {f"{n}_{m}": v for n, sim_res in rmse_result.items() for m, v in sim_res.items()},
    orient='index')

rmse_df['mean_rmse'] = rmse_df.mean(axis=1)
print(rmse_df)
best_params = rmse_df['mean_rmse'].idxmin()
best_f, best_e = best_params.split('_')
print()
print(f'The best number of latent factors is: {best_f} with number of epoch: {best_e} ')


## Step 4
Run the models with the optimal hyperparameters to the whole training
set.


### KNN(K = 6, Measure = pearson)

sim_options = {'name': best_sim,
               'user_based': True
               }
algo_knn = KNNWithMeans(k= 6, 
                    random_state=0,
                    sim_options= sim_options,
                    verbose=False)

trainset = training_matrix.build_full_trainset()# includes the entire dataset for training
algo_knn.fit(trainset)
pred_KNN = algo_knn.test(unobserved_ratings)

average_prediction = round(np.mean([pred.est for pred in pred_KNN if pred.uid in test_users]), 3)

print("Average of predictions: ", average_prediction)

### SVD(latent factor=30， epoch=20)

algo_SVD = SVD(n_factors=30, n_epochs=20)
algo_SVD.fit(trainset)

pred_SVD = algo_SVD.test(unobserved_ratings)

average_preds = round(np.mean([pred.est for pred in pred_SVD if pred.uid in test_users]), 3)

print("Average of predictions:", average_preds)

## Step 5
Use the final models to rank the non-rated items for each user. This
ranking will be used for the evaluation part next week.


#general method for top-k recommendations
from collections import defaultdict
from surprise.prediction_algorithms.predictions import Prediction
from typing import Dict, List
import numpy as np

def get_top_k(predictions: List[Prediction], 
              k: int) -> Dict[str, List]:
    """Compute the top-K recommendation for each user from a set of predictions.
    Args:
        predictions(list of Prediction objects): The list of predictions, as
            returned by the test method of an algorithm.
        k(int): The number of recommendation to output for each user.
    Returns:
        A dict where keys are user (raw) ids and values are lists of tuples:
        [(raw item id, rating estimation), ...] of size n.
    """
    topk = defaultdict(list)

    for pred in predictions:
        uid = pred.uid 
        iid = pred.iid 
        est = pred.est 
        topk[uid].append((iid, est))
    
    # Sort the predictions
    for uid, user_ratings in topk.items():
        user_ratings.sort(key=lambda x: x[1], reverse=True)
        topk[uid] = user_ratings[:k]
    return topk


def print_top_k(user_id: str, topk: Dict[str, List]) -> None:
    user_ratings = topk[user_id]
    print(f"TOP-{len(user_ratings)} predictions for user {user_id}: {[(item, round(rating,2)) for (item, rating) in user_ratings]}")


### KNN

top5_knn = get_top_k(pred_KNN, k=5)
top10_knn = get_top_k(pred_KNN, k=10)
top20_knn = get_top_k(pred_KNN, k=20)

print_top_k("AE23LDQTB7L76AP6E6WPBFVYL5DA", top5_knn)
print_top_k("AE23LDQTB7L76AP6E6WPBFVYL5DA", top10_knn)
print_top_k("AE23LDQTB7L76AP6E6WPBFVYL5DA", top20_knn)

### SVD

top5_svd = get_top_k(pred_SVD, k=5)
top10_svd = get_top_k(pred_SVD, k=10)
top20_svd = get_top_k(pred_SVD, k=20)

print_top_k("AE23LDQTB7L76AP6E6WPBFVYL5DA", top5_svd)
print_top_k("AE23LDQTB7L76AP6E6WPBFVYL5DA", top10_svd)
print_top_k("AE23LDQTB7L76AP6E6WPBFVYL5DA", top20_svd)
